+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.1.2 (2016-09-21)
+
+Features:
+
+  - new `--pytest-option` flag to pass CLI options on to pytest
+  - short (single-character) flags for commonly-used CLI options
+
+Fixes:
+
+  - improved `--help` CLI output
+
+
 ## 1.1.1 (2016-09-20)
 
 Initial implementation of changelog.